requirements.txt
Django==4.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
# Firebase dependency removed
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.115.0
celery==5.3.6
redis==5.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-debug-toolbar==4.3.0
pillow==10.2.0
django-storages==1.14.2
gunicorn==21.2.0
pytest==8.0.0
pytest-django==4.7.0
black==24.2.0
flake8==7.0.0
isort==5.13.2
django-crontab==0.7.1 # Or the version pip installed
python-dateutil==2.8.2 # Or latest compatible version
